chore(repo): close autonomous maintenance follow-ups
